# services/api/docs/swagger.yaml
components:
  schemas:
...
    auth.AccessTokenResponse:
      properties:
        access_token:
          type: string
      type: object
...
    auth.CredentialsAddBeginRequest:
      properties:
        device_name:
          type: string
      type: object
    auth.CredentialsAddCompleteRequest:
      properties:
        credential:
          type: object
        session_key:
          type: string
      type: object
...
    auth.LoginCompleteRequest:
      properties:
        credential:
          type: object
        session_key:
          type: string
      type: object
    auth.LoginCompleteResponse:
      properties:
        access_token:
          type: string
        user:
          $ref: '#/components/schemas/auth.UserSummary'
      type: object
...
    auth.OnboardBeginRequest:
      properties:
        device_name:
          type: string
        invite_id:
          type: string
        token:
          type: string
      type: object
    auth.OnboardBeginResponse:
      properties:
        publicKey:
          type: object
        session_key:
          type: string
        user_id:
          type: string
      type: object
    auth.OnboardCompleteRequest:
      properties:
        credential:
          type: object
        invite_id:
          type: string
        session_key:
          type: string
      type: object
    auth.PublicKeyOptionsResponse:
      properties:
        publicKey:
          type: object
        session_key:
          type: string
      type: object
...
    auth.RecoveryInitRequest:
      properties:
        email:
          type: string
      type: object
    auth.RecoveryInitResponse:
      properties:
        flow_id:
          type: string
      type: object
    auth.RecoveryRegisterBeginRequest:
      properties:
        device_name:
          type: string
        flow_id:
          type: string
      type: object
    auth.RecoveryRegisterCompleteRequest:
      properties:
        credential:
          type: object
        flow_id:
          type: string
        session_key:
          type: string
      type: object
    auth.RecoveryVerifyRequest:
      properties:
        code:
          type: string
        flow_id:
          type: string
      type: object
    auth.RecoveryVerifyResponse:
      properties:
        flow_id:
          type: string
        user_id:
          type: string
      type: object
...
    auth.UserSummary:
      properties:
        email:
          type: string
        full_name:
          type: string
        id:
          type: string
        roles:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
info:
  title: ""
  version: ""
openapi: 3.1.0
paths:
...
  /api/v1/auth/credentials/add/begin:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.CredentialsAddBeginRequest'
        description: Begin credential add
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.PublicKeyOptionsResponse'
          description: OK
      summary: Add credential (begin)
      tags:
      - auth
  /api/v1/auth/credentials/add/complete:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.CredentialsAddCompleteRequest'
        description: Complete credential add
        required: true
      responses:
        "204":
          description: No Content
      summary: Add credential (complete)
      tags:
      - auth
...
  /api/v1/auth/login/complete:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.LoginCompleteRequest'
        description: Login complete request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.LoginCompleteResponse'
          description: OK
      summary: Complete login
      tags:
      - auth
...
  /api/v1/auth/onboard/begin:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.OnboardBeginRequest'
        description: Onboard begin request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.OnboardBeginResponse'
          description: OK
      summary: Onboard begin
      tags:
      - auth
  /api/v1/auth/onboard/complete:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.OnboardCompleteRequest'
        description: Onboard complete request
        required: true
      responses:
        "204":
          description: No Content
      summary: Onboard complete
      tags:
      - auth
...
  /api/v1/auth/recovery/init:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.RecoveryInitRequest'
        description: Recovery init request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.RecoveryInitResponse'
          description: OK
      summary: Recovery init
      tags:
      - auth
  /api/v1/auth/recovery/register/begin:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.RecoveryRegisterBeginRequest'
        description: Recovery register begin request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.PublicKeyOptionsResponse'
          description: OK
      summary: Recovery register begin
      tags:
      - auth
  /api/v1/auth/recovery/register/complete:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.RecoveryRegisterCompleteRequest'
        description: Recovery register complete request
        required: true
      responses:
        "204":
          description: No Content
      summary: Recovery register complete
      tags:
      - auth
  /api/v1/auth/recovery/verify:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth.RecoveryVerifyRequest'
        description: Recovery verify request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.RecoveryVerifyResponse'
          description: OK
      summary: Recovery verify
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/auth.AccessTokenResponse'
          description: OK
      summary: Refresh access token
      tags:
      - auth
...